Add specific database schema blueprint integration classes to rdbms
